src/requirements.txt
matplotlib==3.3.4
numpy==1.19.5
pandas==1.3.3
Pillow==9.4.0
torch==1.13.1
torchvision==0.14.1
